+++ lib/device_info_plus.dart
@@ -1,4 +1,4 @@
 /// Auxillary library for accessing `device_info_plus` directly (if needed)
-library flutter_app_info_device_info_plus;
+library;
 
 export 'package:device_info_plus/device_info_plus.dart';